Fix app crash: Move launch button logic to DashboardFragment
